refactor(cli): extract agent-loop group into commands/agent_loop_commands.py
Move the `agent-loop start` and `agent-loop brief` subcommands out of
cli.py into a dedicated module. Uses lazy import for AgentLoopOrchestrator
so the heavier agent module only loads when the command is invoked.

cli.py shrinks by ~45 lines.
